travel-itinerary-cr-with-backend/backend/requirements.txt
django>=4.2
djangorestframework>=3.15.0
django-cors-headers>=4.0.0
mysqlclient>=2.2.0
